Extract constants file
