+++ src/constants.js
@@ -1,0 +1,6 @@
+export const ENTER_KEY = 'Enter';
+export const SPACEBAR = ' ';
+export const ARROW_DOWN = 'ArrowDown';
+export const ARROW_UP = 'ArrowUp';
+export const HOME_KEY = 'Home';
+export const END_KEY = 'End';
